fix(follow): create FollowDao
