+++ src/main/java/io/github/teamwaff1e/waffle/domain/member/dao/MemberDao.java
@@ -42,22 +42,5 @@ public class MemberDao implements CrudDao<Member, Long> {
         entityManager.remove(member);
     }
 
-    public void unfollow(Long memberId, Long followingId){
 
-        List<Follow> list = entityManager.createQuery("select f from Follow as f where f.memberId=:id and f.followingId=:followingId",Follow.class)
-                .setParameter("id" , memberId)
-                .setParameter("followingId",followingId)
-                .getResultList();
-        System.out.println(list.get(0));
-        entityManager.remove(list.get(0));
-    }
-    public void follow(Follow follow){
-        entityManager.persist(follow);
-    }
-
-    public List<Follow> findFollowById(Long memberId){
-        return entityManager.createQuery("select f from Follow as f where f.memberId=:id",Follow.class)
-                .setParameter("id" , memberId)
-                .getResultList();
-    }
 }
